added delete function for posts

+++ main/views.py
@@ -142,8 +142,12 @@ def post_previews(request):
 
 def edit_post(request, id):
     if request.method == 'DELETE':
-        Post.objects.get(id=id).delete()
-        return HttpResponse(status=204)
+        if is_employee(request.user):
+
+            Post.objects.get(id=id).delete()
+            return HttpResponse(status=204)
+        else:
+            return HttpResponse(status=401)
 
     elif request.method == 'GET':
         post = Post.objects.get(id=id)
